+++ packages/shared/forks/ReactFeatureFlags.persistent.js
@@ -23,6 +23,7 @@ export const enableSchedulerTracing = __PROFILE__;
 export const enableSuspenseServerRenderer = false;
 export const disableInputAttributeSyncing = false;
 export const enableStableConcurrentModeAPIs = false;
+export const warnAboutShorthandPropertyCollision = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
